refactor: move examples to ranim-examples crate
