+++ Cargo.toml
@@ -138,64 +138,3 @@ pre-release-hook = [
 pre-release-replacements = [
   { file = "README.md", search = "ranim = \"[^\"]+\"", replace = "{{crate_name}} = \"{{version}}\"" },
 ]
-
-# MARK: Examples
-[[example]]
-name = "arc"
-path = "examples/arc/main.rs"
-
-[package.metadata.example.arc]
-wasm = true
-
-[[example]]
-name = "arc_between_points"
-path = "examples/arc_between_points/main.rs"
-
-[package.metadata.example.arc_between_points]
-wasm = true
-
-[[example]]
-name = "basic"
-path = "examples/basic/main.rs"
-
-[package.metadata.example.basic]
-wasm = true
-
-[[example]]
-name = "hanoi"
-path = "examples/hanoi/main.rs"
-
-[package.metadata.example.hanoi]
-wasm = true
-
-[[example]]
-name = "hello_ranim"
-path = "examples/hello_ranim/main.rs"
-
-[package.metadata.example.hello_ranim]
-wasm = true
-
-[[example]]
-name = "palettes"
-path = "examples/palettes/main.rs"
-
-[[example]]
-name = "perspective_blend"
-path = "examples/perspective_blend/main.rs"
-
-[package.metadata.example.perspective_blend]
-wasm = true
-
-[[example]]
-name = "ranim_logo"
-path = "examples/ranim_logo/main.rs"
-
-[package.metadata.example.ranim_logo]
-wasm = true
-
-[[example]]
-name = "extract_vitem_visualize"
-path = "examples/extract_vitem_visualize/main.rs"
-
-[package.metadata.example.extract_vitem_visualize]
-wasm = true
