--- examples/arc/README.md
@@ -1,8 +1,0 @@
-# Arc
-
-This example demonstrates `Arc`.
-
-https://github.com/user-attachments/assets/4f8c6f36-e6b7-429b-8d2a-91c21e17dec2
-
-> [arc.mp4](../../assets/arc.mp4)
-
