--- examples/arc_between_points/README.md
@@ -1,7 +1,0 @@
-# ArcBetweenPoints Example
-
-This example demonstrates `ArcBetweenPoints`.
-
-https://github.com/user-attachments/assets/9595ce1d-2390-4d4d-958f-9299755460b7
-
-> [arc_between_points.mp4](../../assets/arc_between_points.mp4)
