--- examples/basic/README.md
@@ -1,7 +1,0 @@
-# Basic Example
-
-This example demonstrates the basic usage of `ranim`.
-
-https://github.com/user-attachments/assets/dadab1ab-65e4-4cd3-bda4-b2ef97c61b0b
-
-> [basic.mp4](../../assets/basic.mp4)
