--- examples/palettes/README.md
@@ -1,5 +1,0 @@
-# Palettes
-
-Palettes from manim
-
-![palette](../../assets/palettes.png)
