+++ packages/ranim-examples/Cargo.toml
@@ -17,6 +17,7 @@ crate-type = ["rlib", "cdylib"] # use cdylib to produce .wasm file
 [dependencies]
 ranim.workspace = true
 ranim-macros.workspace = true
+itertools.workspace = true
 rand = "0.9.2"
 rand_chacha = "0.9.0"
 wasm-bindgen = "0.2.100"
