+++ packages/ranim-examples/src/vitem/svg.rs
@@ -1,9 +1,8 @@
-use glam::DVec3;
-use log::LevelFilter;
 use ranim::{
     animation::{creation::WritingAnim, fading::FadingAnim, lagged::LaggedAnim},
     color::palettes::manim,
     components::ScaleHint,
+    glam::DVec3,
     items::{
         Group,
         vitem::{VItem, svg::SvgItem, typst::typst_svg},
@@ -12,12 +11,14 @@ use ranim::{
     timeline::TimeMark,
 };
 
-const SVG: &str = include_str!("../../assets/Ghostscript_Tiger.svg");
+const SVG: &str = include_str!("../../../../assets/Ghostscript_Tiger.svg");
 
 #[scene]
+#[wasm_demo_doc]
 #[preview]
 #[output(dir = "basic")]
-fn basic(r: &mut RanimScene) {
+/// An example shows [`SvgItem`]
+fn svg_item(r: &mut RanimScene) {
     let _r_cam = r.insert_and_show(CameraFrame::default());
     r.timelines_mut().forward(0.2);
 
@@ -55,23 +56,4 @@ fn basic(r: &mut RanimScene) {
         r.timelines().max_total_secs(),
         TimeMark::Capture("preview.png".to_string()),
     );
-}
-
-fn main() {
-    #[cfg(not(target_arch = "wasm32"))]
-    {
-        #[cfg(debug_assertions)]
-        pretty_env_logger::formatted_timed_builder()
-            .filter(Some("ranim"), LevelFilter::Trace)
-            .init();
-        #[cfg(not(debug_assertions))]
-        pretty_env_logger::formatted_timed_builder()
-            .filter(Some("ranim"), LevelFilter::Info)
-            .init();
-    }
-
-    #[cfg(feature = "app")]
-    preview(basic_scene);
-    #[cfg(not(feature = "app"))]
-    render_scene(basic_scene)
 }
