+++ packages/ranim-macros/src/lib.rs
@@ -61,6 +61,13 @@ pub fn scene(args: TokenStream, input: TokenStream) -> TokenStream {
         .filter(|attr| attr.path().is_ident("doc"))
         .collect();
 
+    let doc_head = doc_attrs.first();
+    let doc_attrs = if doc_attrs.len() > 1 {
+        &doc_attrs[1..]
+    } else {
+        &[]
+    };
+
     // 场景名称
     let scene_name = attrs.name.unwrap_or_else(|| fn_name.to_string());
 
@@ -99,17 +106,27 @@ pub fn scene(args: TokenStream, input: TokenStream) -> TokenStream {
     }
 
     let preview = attrs.preview;
-    let doc = if attrs.wasm_demo_doc {
-        quote! {
-            #[doc = concat!("<canvas id=\"ranim-app-", stringify!(#fn_name), "\" width=\"1280\" height=\"720\" style=\"width: 100%;\"></canvas>")]
-            #[doc = concat!("<script type=\"module\">")]
-            #[doc = concat!("  const { find_scene } = await ranim_examples;")]
-            #[doc = concat!("  find_scene(\"", stringify!(#fn_name), "\").run_app();")]
-            #[doc = "</script>"]
-        }
-    } else {
-        quote! {}
+    let mut doc = quote! {
+        #(#doc_attrs)*
     };
+    if attrs.wasm_demo_doc {
+        doc = quote! {
+            #[doc = concat!(r#"<div>"#)]
+            #[doc = concat!(r#"  <canvas id="ranim-app-"#, stringify!(#fn_name), r#"" width="1280" height="720" style="width: 100%; height: 100%;"></canvas>"#)]
+            #[doc = concat!(r#"  <script type="module">"#)]
+            #[doc = concat!(r#"    const { find_scene } = await ranim_examples;"#)]
+            #[doc = concat!(r#"    find_scene(""#, stringify!(#fn_name), r#"").run_app();"#)]
+            #[doc = concat!(r#"  </script>"#)]
+            #[doc = concat!(r#"</div>"#)]
+            #doc
+        };
+    }
+    if let Some(doc_head) = doc_head {
+        doc = quote! {
+            #doc_head
+            #doc
+        };
+    }
 
     let static_output_name = syn::Ident::new(
         &format!("__SCENE_{}_OUTPUTS", fn_name.to_string().to_uppercase()),
@@ -136,7 +153,6 @@ pub fn scene(args: TokenStream, input: TokenStream) -> TokenStream {
     // 构造 Scene 并塞进分布式切片
     let expanded = quote! {
         #doc
-        #(#doc_attrs)*
         #vis fn #fn_name(r: &mut #ranim::timeline::RanimScene) #fn_body
 
         #[doc(hidden)]
